Reference

HypertextLiteral.HypertextLiteral — Module

HypertextLiteral

This library provides for a @htl() macro and a htl string literal, both implementing interpolation that is aware of hypertext escape context. The @htl macro has the advantage of using Julia's native string parsing, so that it can handle arbitrarily deep nesting. However, it is a more verbose than the htl string literal and doesn't permit interpolated string literals. Conversely, the htl string literal, @htl_str, uses custom parsing letting it handle string literal escaping, however, it can only be used two levels deep (using three quotes for the outer nesting, and a single double quote for the inner).

HypertextLiteral.EscapeProxy — Type

EscapeProxy(io) - wrap an `io` to perform HTML escaping

This is a transparent proxy that performs HTML escaping so that objects that are printed are properly converted into valid HTML values. As a special case, HTML objects are not escaped, hence, they could be used to bypass this proxy. At this time, it only supports & and < escaping.

Examples

julia> ep = EscapeProxy(stdout);
julia> print(ep, "A&B")
A&amp;B
julia> print(ep, HTML("<tag/>"))
<tag/>

HypertextLiteral.Result — Type

Result(expr, unwrap)

Address display modalities by showing the macro expression that generated the results when shown on the REPL. However, when used with print() show the results. This object is also showable to any IO stream via "text/html".

HypertextLiteral.UnwrapHTML — Type

UnwrapHTML(data) - delegate regular printing to text/html

This is the inverse wrapper to Docs.HTML – instead of enabling regular objects to be printed as MIME"text/html" it unwraps an HTML or any other object showable as text/html to be printable. Conversely, it will cause an error if that object is not showable as text/html.

Examples

julia> print(UnwrapHTML(HTML("Hello World")))
Hello World

HypertextLiteral.attribute_pair — Method

attribute_pair(name, value)

Wrap and escape attribute name and pair within a single-quoted context so that it is showable("text/html"). It's assumed that the attribute name has already been normalized.

If an attribute value is Bool or Nothing, then special treatment is provided. If the value is false or nothing then the entire pair is not printed. If the value is true than an empty string is produced.

HypertextLiteral.attributes — Method

attributes(value)

Convert Julian object into a serialization of attribute pairs, showable via MIME"text/html". The default implementation of this delegates value construction of each pair to attribute_pair().

HypertextLiteral.content — Method

content(value)

Wrap content so that it is showable("text/html"). By default, string values are escaped with escape_content. For performance reasons, we reproduce Number and Symbol objects without escaping them. Tuples, arrays and generators are represented by concatenating their elements. As a fallback, we assume the value has implemented show() for MIME"text/html", if not, a MethodError will result.

HypertextLiteral.double_quoted — Method

double_quoted(value)

Wrap and escape a double-quoted attribute value so that it is showable("text/html"). This uses stringify to do the actual conversion of the attribute to a usable string value.

HypertextLiteral.interpolate — Method

interpolate(args, this)::Expr

Take an interweaved set of Julia expressions and strings, tokenize the strings according to the HTML specification [1], wrapping the expressions with wrappers based upon the escaping context, and returning an expression that combines the result with an Result wrapper.

For these purposes, a Symbol is treated as an expression to be resolved; while a String is treated as a literal string that won't be escaped. Critically, interpolated strings to be escaped are represented as an Expr with head of :string.

There are tags, "script" and "style" which are rawtext, in these cases there is no escaping, and instead raise an exception if the appropriate ending tag is in substituted content.

[1] https://html.spec.whatwg.org/multipage/parsing.html#tokenization

HypertextLiteral.interpolate_attributes — Method

interpolate_attributes(element, expr)::Vector{Expr}

Continue conversion of an arbitrary Julia expression within the attribute section of the given element.

HypertextLiteral.normalize_attribute_name — Method

normalize_attribute_name(name)::String

For String names, this simply verifies that they pass the attribute name production, but are otherwise untouched.

For Symbol names, this converts snake_case Symbol objects to their kebab-case equivalent. So that keywords, such as for could be used, we strip leading underscores.

HypertextLiteral.rawtext — Method

rawtext(context, value)

Wrap a string value that occurs with RAWTEXT, SCRIPT and other element context so that it is showable("text/html"). The default implementation ensures that the given value doesn't contain substrings illegal for the given context.

HypertextLiteral.single_quoted — Method

single_quoted(value)

Wrap and escape a single-quoted attribute value so that it is showable("text/html"). This uses stringify to do the actual conversion of the attribute to a usable string value.

HypertextLiteral.stringify — Method

stringify(value)

Convert an attribute valueto aString`. The value returned will then be escaped depending upon the particular context.

String values are returned as-is

Number and Symbol values are converted to a String

There is no general fallback, hence, a MethodError will result when attempting to stringify most data types. If your application would like to stringify all attribute values, you could register this fallback.

HypertextLiteral.stringify(value) = string(value)

HypertextLiteral.@htl — Macro

@htl string-expression

Create a Result object with string interpolation ($) that uses context-sensitive hypertext escaping. Before Julia 1.6, interpolated string literals, e.g. $("Strunk & White"), are treated as errors since they cannot be reliably detected (see Julia issue #38501).

HypertextLiteral.@htl_str — Macro

@htl_str -> Result

Create a Result object with string interpolation ($) that uses context-sensitive hypertext escaping. Escape sequences should work identically to Julia strings, except in cases where a slash immediately precedes the double quote (see @raw_str and Julia issue #22926).

Interpolation is extended beyond regular Julia strings to handle three additional cases: tuples, named tuples (for attributes), and generators. See Julia #38734 for the feature request so that this could also work within the @htl macro syntax.