Reference

HypertextLiteral.HypertextLiteral — Module

HypertextLiteral

This library provides for a @htl() macro and a htl string literal, both implementing interpolation that is aware of hypertext escape context. The @htl macro has the advantage of using Julia's native string parsing, so that it can handle arbitrarily deep nesting. However, it is a more verbose than the htl string literal and doesn't permit interpolated string literals. Conversely, the htl string literal, @htl_str, uses custom parsing letting it handle string literal escaping, however, it can only be used two levels deep (using three quotes for the outer nesting, and a single double quote for the inner).

HypertextLiteral.Attribute — Type

Attribute{name}

This parameterized type to represent HTML attributes so that we could dispatch serialization of custom attributes and data types. This is modeled upon the MIME data type. Values written in this way must be escaped for single-quoted context (' => "&apos;", & => "&amp;").

HypertextLiteral.Result — Type

Result(expr, xs...)

Create an object that is showable to "text/html" created from arguments that are also showable. Leaf entries can be created using HTML. This expression additionally has an expression which is used when displaying the object to the REPL. Calling print will produce rendered output.

HypertextLiteral.attribute_pair — Method

attribute_pair(attribute, value)

Wrap and escape attribute name and pair within a single-quoted context so that it is showable("text/html"). This uses stringify to do the actual conversion of the attribute to a usable string value. If an attribute is_boolean it is given special treatment, for true values, the attribute is printed with an empty string, else it is omitted. Moreover, attributes with value of nothing are coalesced to the empty string (unless they are boolean, in which case they are omitted).

HypertextLiteral.attributes — Method

attributes(element::Symbol, value)

Convert Julian object into a serialization of attribute pairs, showable via MIME"text/html". The default implementation of this delegates value construction of each pair to attribute_pair().

HypertextLiteral.content — Method

content(value)

Wrap and escape content so that it is showable("text/html"). By default, we handle strings, numbers and symbols by escaping them. Tuples, arrays and generators are wrapped by concatenating their elements. As a fallback, we assume the value has implemented show() for MIME"text/html", if not, a MethodError will result.

HypertextLiteral.css_value — Method

css_value(val)

Convert a native Julia object into a string suitable for use as a CSS value. This is useful for adding support for cssunits or other tools that build CSS fragments.

HypertextLiteral.double_quoted — Method

double_quoted(attribute, value)

Wrap and escape a double-quoted attribute value so that it is showable("text/html"). This uses stringify to do the actual conversion of the attribute to a usable string value.

HypertextLiteral.escape_content — Method

escape_content(value)

Escape a string value for use within HTML content, this includes replacing & with &amp; and < with &lt;. We're not further escaping quotes within content since benchmarking shows us that it adds about 10% on the runtime for each character escaped.

HypertextLiteral.htl_escape_value — Method

htl_escape_value(s)

Perform extensive escaping needed for a string to be used as an unquoted attribute. This can also be used for quoted values or within element content (although it's overkill in those contexts).

HypertextLiteral.interpolate — Method

interpolate(args, this)::Expr

Take an interweaved set of Julia expressions and strings, tokenize the strings according to the HTML specification [1], wrapping the expressions with wrappers based upon the escaping context, and returning an expression that combines the result with an Result wrapper.

For these purposes, a Symbol is treated as an expression to be resolved; while a String is treated as a literal string that won't be escaped. Critically, interpolated strings to be escaped are represented as an Expr with head of :string.

There are tags, "script" and "style" which are rawtext, in these cases there is no escaping, and instead raise an exception if the appropriate ending tag is in substituted content.

[1] https://html.spec.whatwg.org/multipage/parsing.html#tokenization

HypertextLiteral.interpolate_attributes — Method

interpolate_attributes(element, expr)::Vector{Expr}

Continue conversion of an arbitrary Julia expression within the attribute section of the given element.

HypertextLiteral.is_boolean — Method

is_boolean(attribute::Attribute)

This function returns true if the given attribute is boolean. In in such case, false or nothing means the attribute should be removed from the produced output. Note that there are some HTML attributes which may take a boolean value but that produce on or off, or something else. Those attributes are not considered boolean.

HypertextLiteral.kebab_case — Method

kebab_case(s)

This converts an name in PascalCase, camelCase, and snake_case into their kebab-case equivalent. This will lowercase the name. This transformation is problematic for SVG attribute names, which need to be in camelCase.

HypertextLiteral.rawtext — Method

rawtext(context, value)

Wrap a string value that occurs with RAWTEXT, SCRIPT and other element context so that it is showable("text/html"). The default implementation ensures that the given value doesn't contain substrings illegal for the given context.

HypertextLiteral.single_quoted — Method

single_quoted(attribute, value)

Wrap and escape a single-quoted attribute value so that it is showable("text/html"). This uses stringify to do the actual conversion of the attribute to a usable string value.

HypertextLiteral.stringify — Method

stringify(attribute::Attribute, value)

Convert an attribute valueto aString` suitable for inclusion into the given attribute's value. The value returned will then be escaped depending upon the particular context, single/double or unquoted.

String values are returned as-is

Number and Symbol values are converted to a String

Bool values of known boolean attributes produce an error.

Nothing becomes an empty string (unless for boolean attribute).

There is no general fallback, hence, a MethodError will result when attempting to stringify most data types. If your application would like to stringify all attribute values, you could register this fallback.

stringify(::Attribute, value) = string(value)

HypertextLiteral.stringify — Method

stringify(a::Attribute{name})

This provides the serialization of a given normalized attribute so that camelCase could be preserved on output for elements foreign to HTML, such as SVG. By default, about 2 dozen SVG attributes are defined.

HypertextLiteral.@htl — Macro

@htl string-expression

Create a Result object with string interpolation ($) that uses context-sensitive hypertext escaping. Before Julia 1.6, interpolated string literals, e.g. $("Strunk & White"), are treated as errors since they cannot be reliably detected (see Julia issue #38501).

HypertextLiteral.@htl_str — Macro

@htl_str -> Result

Create a Result object with string interpolation ($) that uses context-sensitive hypertext escaping. Escape sequences should work identically to Julia strings, except in cases where a slash immediately precedes the double quote (see @raw_str and Julia issue #22926).

Interpolation is extended beyond regular Julia strings to handle three additional cases: tuples, named tuples (for attributes), and generators. See Julia #38734 for the feature request so that this could also work within the @htl macro syntax.