Reference

HypertextLiteral.HypertextLiteral — Module

HypertextLiteral

This library provides for a @htl() macro and a htl string literal, both implementing interpolation that is aware of hypertext escape if hasmethod(show, Tuple{IO, typeof(mime), eltype(value)}) for item in value show(io, mime, item) end return end context. The @htl macro has the advantage of using Julia's native string parsing, so that it can handle arbitrarily deep nesting. However, it is a more verbose than the htl string literal and doesn't permit interpolated string literals. Conversely, the htl string literal, @htl_str, uses custom parsing letting it handle string literal escaping, however, it can only be used two levels deep (using three quotes for the outer nesting, and a single double quote for the inner).

HypertextLiteral.HTL — Type

HTL(s): Create an array of objects that render as html.

HTL("<div>foo</div>")

This is similar HTML{Vector} with a few exceptions. First, the contents of the vector are concatenated. Second, direct rendering is limited to AbstractString, others are delegated to show. Third, the splat constructor converts arguments to an HTL vector.

Finally, regular display of the value to the terminal renders the objects and produces the equivalent string representation.

HypertextLiteral.HTLAttribute — Type

HTLAttribute{name}

This parameterized type to represent HTML attributes so that we could dispatch serialization of custom attributes and data types. This is modeled upon the MIME data type. Values written in this way must be escaped for use in an unquoted string, htl_escape_value can do this.

HypertextLiteral.InterpolatedValue — Type

Interpolated Value

This abstract type represents a value that must be escaped. The various subclasses provide the context for the escaping. They include:

ElementData Values expanded as element content, including text nodes and/or subordinate HTML fragments ElementAttributes Values to be expanded as attribute/value pairs AttributePair Unquoted name/value pair for attributes; handles special cases of boolean attributes AttributeDoubleQuoted Value serialized within double quoted attribute AttributeSingleQuoted Value serialized within single quoted attribute

The string interpolation is here is conservative. To provide express data type conversions for ElementData, override show "text/html".

HypertextLiteral.css_value — Method

css_value(val)

Convert a native Julia object into a string suitable for use as a CSS value. This is useful for adding support for cssunits or other tools that build CSS fragments.

HypertextLiteral.htl_escape_value — Method

htl_escape_value(s)

Perform extensive escaping needed for a string to be used as an unquoted attribute. This can also be used for quoted values or within element content (although it's overkill in those contexts).

HypertextLiteral.htl_is_boolean — Method

htl_is_boolean(attribute::HTLAttribute)

This function returns true if the given attribute is boolean. In in such case, false or nothing means the attribute should be removed from the produced output. Note that there are some HTML attributes which may take a boolean value but that produce on or off, or something else. Those attributes are not considered boolean.

HypertextLiteral.htl_normalize — Method

htl_normalize(s)

This converts an name using either camelCase or unix_case names into their dashed equivalents. This will lowercase the name.

HypertextLiteral.htl_represent — Method

htl_represent(a::HTLAttribute{name})

This provides the serialization of a given normalized attribute so that camelCase could be preserved on output for elements foreign to HTML, such as SVG. By default, about 2 dozen SVG attributes are defined.

HypertextLiteral.htl_stringify — Method

htl_stringify(attribute::HTLAttribute, value)

Convert an attribute valueto aString` suitable for inclusion into the given attribute's value. The value returned will then be escaped depending upon the particular context, single/double or unquoted.

String values are returned as-is

Number and Symbol values are converted to a String

Bool values of known boolean attributes produce an error.

There is no general fallback, hence, a MethodError will result when attempting to stringify most data types. If your application would like to stringify all attribute values, you could register this fallback.

htl_stringify(::HTLAttribute, value) = string(value)

HypertextLiteral.interpolate — Method

interpolate(args):Expr

Take an interweaved set of Julia expressions and strings, tokenize the strings according to the HTML specification [1], wrapping the expressions with wrappers based upon the escaping context, and returning an expression that combines the result with an HTL wrapper.

For these purposes, a Symbol is treated as an expression to be resolved; while a String is treated as a literal string that won't be escaped. Critically, interpolated strings to be escaped are represented as an Expr with head of :string.

There are tags, "script" and "style" which are rawtext, in these cases there is no escaping, and instead raise an exception if the appropriate ending tag is in substituted content.

[1] https://html.spec.whatwg.org/multipage/parsing.html#tokenization

HypertextLiteral.@htl — Macro

@htl string-expression

Create a HTL object with string interpolation ($) that uses context-sensitive hypertext escaping. Before Julia 1.6, interpolated string literals, e.g. $("Strunk & White"), are treated as errors since they cannot be reliably detected (see Julia issue #38501).

HypertextLiteral.@htl_str — Macro

@htl_str -> HTL

Create a HTL object with string interpolation ($) that uses context-sensitive hypertext escaping. Escape sequences should work identically to Julia strings, except in cases where a slash immediately precedes the double quote (see @raw_str and Julia issue #22926).

Interpolation is extended beyond regular Julia strings to handle three additional cases: tuples, named tuples (for attributes), and generators. See Julia #38734 for the feature request so that this could also work within the @htl macro syntax.