Reference

HypertextLiteral.HypertextLiteral — Module

HypertextLiteral

This library provides for a @htl() macro and a htl string literal, both implementing interpolation that is aware of hypertext escape context. The @htl macro has the advantage of using Julia's native string parsing, so that it can handle arbitrarily deep nesting. However, it is a more verbose than the htl string literal and doesn't permit interpolated string literals. Conversely, the htl string literal, @htl_str, uses custom parsing letting it handle string literal escaping, however, it can only be used two levels deep (using three quotes for the outer nesting, and a single double quote for the inner).

HypertextLiteral.EscapeProxy — Type

EscapeProxy(io) - wrap an `io` to perform HTML escaping

This is a transparent proxy that performs HTML escaping so that objects that are printed are properly converted into valid HTML values. As a special case, objects wrapped with Bypass are not escaped, and bypass the proxy.

Examples

julia> ep = EscapeProxy(stdout);
julia> print(ep, "A&B")
A&amp;B
julia> print(ep, Bypass("<tag/>"))
<tag/>

HypertextLiteral.Render — Type

Render(data) - printed object shows its text/html

HypertextLiteral.Reprint — Type

Reprint(fn) -- apply the lambda function when printed

HypertextLiteral.Result — Type

Result(expr, unwrap)

Address display modalities by showing the macro expression that generated the results when shown on the REPL. However, when used with print() show the results. This object is also showable to any IO stream via "text/html".

HypertextLiteral.attribute_value — Method

attribute_value(x)

This method may be implemented to specify a printed representation suitable for use within a quoted attribute value. By default, the print representation of an object is used, and then propertly escaped. There are a few overrides that we provide.

The elements of a Tuple or AbstractArray object are printed, with a space between each item.

The Pair, NamedTuple, and Dict objects are treated as if they are CSS style elements, with a colon between key and value, each pair delimited by a semi-colon.

The Bool object, which has special treatment for bare inside_tag, is an error when used within a quoted attribute.

If an object is wrapped with HTML then it is included in the quoted attribute value as-is, without inspection or escaping.

HypertextLiteral.content — Method

content(x)

This method may be implemented to specify a printed representation suitable for text/html output. As a special case, if the result is wrapped with HTML, then it is passed along as-is. Otherwise, the print representation of the resulting value is escaped. By default AbstractString, Number and Symbol values are printed and escaped. The elements of Tuple and AbstractArray are concatinated and then escaped. If a method is not implemented for a given object, then we attempt to show it via MIME"text/html".

HypertextLiteral.inside_tag — Method

inside_tag(value)

Convert Julian object into a serialization of attribute pairs, showable via MIME"text/html". The default implementation of this delegates value construction of each pair to attribute_pair().

HypertextLiteral.interpolate — Method

interpolate(args, this)::Expr

Take an interweaved set of Julia expressions and strings, tokenize the strings according to the HTML specification [1], wrapping the expressions with wrappers based upon the escaping context, and returning an expression that combines the result with an Result wrapper.

For these purposes, a Symbol is treated as an expression to be resolved; while a String is treated as a literal string that won't be escaped. Critically, interpolated strings to be escaped are represented as an Expr with head of :string.

There are tags, "script" and "style" which are rawtext, in these cases there is no escaping, and instead raise an exception if the appropriate ending tag is in substituted content.

[1] https://html.spec.whatwg.org/multipage/parsing.html#tokenization

HypertextLiteral.normalize_attribute_name — Method

normalize_attribute_name(name)::String

For String names, this simply verifies that they pass the attribute name production, but are otherwise untouched.

For Symbol names, this converts snake_case Symbol objects to their kebab-case equivalent. So that keywords, such as for could be used, we strip leading underscores.

HypertextLiteral.rawtext — Method

rawtext(context, value)

Wrap a string value that occurs with RAWTEXT, SCRIPT and other element context so that it is showable("text/html"). The default implementation ensures that the given value doesn't contain substrings illegal for the given context.

HypertextLiteral.rewrite_inside_tag — Method

rewrite_inside_tag(expr)

Attempt to speed up serialization of inside_tag by exploring the expression tree at macro expansion time.

HypertextLiteral.@htl — Macro

@htl string-expression

Create a Result object with string interpolation ($) that uses context-sensitive hypertext escaping. Before Julia 1.6, interpolated string literals, e.g. $("Strunk & White"), are treated as errors since they cannot be reliably detected (see Julia issue #38501).

HypertextLiteral.@htl_str — Macro

@htl_str -> Result

Create a Result object with string interpolation ($) that uses context-sensitive hypertext escaping. Unlike the @htl macro, this string literal does not include escaping feature [1]. To include $ within user content one must write &#36;. Observe that &quot; and any other HTML ampersand escape sequence can be used as appropriate.

In this syntax, interpolation is extended beyond regular Julia strings to handle three additional cases: tuples, named tuples (for attributes), and generators. See Julia #38734 for the feature request so that this could also work within the @htl macro syntax.

[1] There are also a few edge cases, see @raw_str documentation and Julia #22926 for more detail. See Julia #38948 for a feature request that would provide a paired unicode syntax.